Optimize temporary file.

+++ src/Adapter.php
@@ -37,9 +37,9 @@ class Adapter extends AbstractAdapter
         Conf::setSecretKey($config['secret_key']);
 
         $this->bucket = $config['bucket'];
-        $this->debug = $config['debug'];
+        $this->debug  = $config['debug'];
 
-        $this->setPathPrefix($config['protocol'].'://'.$config['domain'].'/');
+        $this->setPathPrefix($config['protocol'] . '://' . $config['domain'] . '/');
 
         Cosapi::setTimeout($config['timeout']);
     }
@@ -75,24 +75,16 @@ class Adapter extends AbstractAdapter
     {
         $temporaryPath = $this->createTemporaryFile($contents);
 
-        if (false === $temporaryPath) {
+        $response = Cosapi::upload($this->getBucket(), $temporaryPath, $path,
+            null, null, $config->get('insertOnly', 1));
+
+        $response = $this->normalizeResponse($response);
+
+        if (false === $response) {
             return false;
         }
 
-        try {
-            $response = Cosapi::upload($this->getBucket(), $temporaryPath, $path,
-                                        null, null, $config->get('insertOnly', 1));
-
-            $response = $this->normalizeResponse($response);
-
-            if (false === $response) {
-                return false;
-            }
-
-            $this->setContentType($path, $contents);
-        } catch (RuntimeException $exception) {
-            throw $exception;
-        }
+        $this->setContentType($path, $contents);
 
         return $response;
     }
@@ -111,7 +103,7 @@ class Adapter extends AbstractAdapter
         $uri = stream_get_meta_data($resource)['uri'];
 
         $response = Cosapi::upload($this->getBucket(), $uri, $path,
-                                    null, null, $config->get('insertOnly', 1));
+            null, null, $config->get('insertOnly', 1));
 
         $response = $this->normalizeResponse($response);
 
@@ -137,24 +129,16 @@ class Adapter extends AbstractAdapter
     {
         $temporaryPath = $this->createTemporaryFile($contents);
 
-        if (false === $temporaryPath) {
+        $response = Cosapi::upload($this->getBucket(), $temporaryPath, $path,
+            null, null, $config->get('insertOnly', 0));
+
+        $response = $this->normalizeResponse($response);
+
+        if (false === $response) {
             return false;
         }
 
-        try {
-            $response = Cosapi::upload($this->getBucket(), $temporaryPath, $path,
-                                        null, null, $config->get('insertOnly', 0));
-
-            $response = $this->normalizeResponse($response);
-
-            if (false === $response) {
-                return false;
-            }
-
-            $this->setContentType($path, $contents);
-        } catch (RuntimeException $exception) {
-            throw $exception;
-        }
+        $this->setContentType($path, $contents);
 
         return $response;
     }
@@ -173,7 +157,7 @@ class Adapter extends AbstractAdapter
         $uri = stream_get_meta_data($resource)['uri'];
 
         $response = Cosapi::upload($this->getBucket(), $uri, $path,
-                                    null, null, $config->get('insertOnly', 0));
+            null, null, $config->get('insertOnly', 0));
 
         $response = $this->normalizeResponse($response);
 
@@ -194,7 +178,7 @@ class Adapter extends AbstractAdapter
      */
     public function rename($path, $newpath)
     {
-        return (bool) $this->normalizeResponse(
+        return (bool)$this->normalizeResponse(
             Cosapi::move($this->getBucket(), $path, $newpath, 1)
         );
     }
@@ -209,11 +193,8 @@ class Adapter extends AbstractAdapter
     {
         $resource = $this->read($path);
 
-        if (false === $resource) {
-            return false;
-        }
-
-        return (bool) $this->update($newpath, $resource['contents'], new Config());
+        return isset($resource['contents'])
+            ? (bool)$this->update($newpath, $resource['contents'], new Config()) : false;
     }
 
     /**
@@ -223,7 +204,7 @@ class Adapter extends AbstractAdapter
      */
     public function delete($path)
     {
-        return (bool) $this->normalizeResponse(
+        return (bool)$this->normalizeResponse(
             Cosapi::delFile($this->getBucket(), $path)
         );
     }
@@ -235,7 +216,7 @@ class Adapter extends AbstractAdapter
      */
     public function deleteDir($dirname)
     {
-        return (bool) $this->normalizeResponse(
+        return (bool)$this->normalizeResponse(
             Cosapi::delFolder($this->getBucket(), $dirname)
         );
     }
@@ -261,9 +242,10 @@ class Adapter extends AbstractAdapter
      */
     public function setVisibility($path, $visibility)
     {
-        $visibility = $visibility === AdapterInterface::VISIBILITY_PUBLIC ? 'eWPrivateRPublic' : 'eWRPrivate';
+        $visibility = ($visibility === AdapterInterface::VISIBILITY_PUBLIC)
+            ? 'eWPrivateRPublic' : 'eWRPrivate';
 
-        return (bool) $this->normalizeResponse(
+        return (bool)$this->normalizeResponse(
             Cosapi::update($this->getBucket(), $path, null, $visibility)
         );
     }
@@ -276,7 +258,7 @@ class Adapter extends AbstractAdapter
     public function has($path)
     {
         try {
-            return (bool) $this->getMetadata($path);
+            return (bool)$this->getMetadata($path);
         } catch (RuntimeException $exception) {
             return false;
         }
@@ -285,21 +267,25 @@ class Adapter extends AbstractAdapter
     /**
      * @param string $path
      *
-     * @return string
+     * @return array|bool
      */
     public function read($path)
     {
-        return ['contents' => file_get_contents($this->getUrl($path))];
+        $contents = file_get_contents($this->applyPathPrefix($path));
+
+        return $contents !== false ? compact('contents') : false;
     }
 
     /**
      * @param string $path
      *
-     * @return array
+     * @return array|bool
      */
     public function readStream($path)
     {
-        return ['stream' => fopen($this->getUrl($path), 'r')];
+        $stream = fopen($this->applyPathPrefix($path), 'r');
+
+        return $stream !== false ? compact('stream') : false;
     }
 
     /**
@@ -336,11 +322,7 @@ class Adapter extends AbstractAdapter
     {
         $stat = $this->getMetadata($path);
 
-        if (isset($stat['filesize'])) {
-            return ['size' => $stat['filesize']];
-        }
-
-        return false;
+        return isset($stat['filesize']) ? ['size' => $stat['filesize']] : false;
     }
 
     /**
@@ -352,11 +334,8 @@ class Adapter extends AbstractAdapter
     {
         $stat = $this->getMetadata($path);
 
-        if (isset($stat['custom_headers']['Content-Type'])) {
-            return ['mimetype' => $stat['custom_headers']['Content-Type']];
-        }
-
-        return false;
+        return isset($stat['custom_headers']['Content-Type'])
+            ? ['mimetype' => $stat['custom_headers']['Content-Type']] : false;
     }
 
     /**
@@ -368,11 +347,7 @@ class Adapter extends AbstractAdapter
     {
         $stat = $this->getMetadata($path);
 
-        if (isset($stat['ctime'])) {
-            return ['timestamp' => $stat['ctime']];
-        }
-
-        return false;
+        return isset($stat['ctime']) ? ['timestamp' => $stat['ctime']] : false;
     }
 
     /**
@@ -396,37 +371,40 @@ class Adapter extends AbstractAdapter
     }
 
     /**
-	 * Creates a temporary file
-	 *
+     * Creates a temporary file
+     *
      * @param string $content
      *
-     * @return bool|string
+     * @return string
+     * @throws RuntimeException
      */
     protected function createTemporaryFile($content)
     {
         $temporaryPath = $this->getTemporaryPath();
 
-        chmod($temporaryPath, 0777);
+        if (false === $temporaryPath) {
+            throw new RuntimeException("Unable to create temporary file in '{$temporaryPath}'.");
+        }
 
         file_put_contents($temporaryPath, $content);
-		
-		// The file is automatically removed when closed, or when the script ends.
-		register_shutdown_function(function() use($temporaryPath) {
-			unlink($temporaryPath);
-		});
+
+        // The file is automatically removed when closed, or when the script ends.
+        register_shutdown_function(function () use ($temporaryPath) {
+            unlink($temporaryPath);
+        });
 
         return $temporaryPath;
     }
-	
+
     /**
      * Gets a temporary file path.
      *
-     * @return string
+     * @return bool|string
      */
     protected function getTemporaryPath()
     {
         return tempnam(sys_get_temp_dir(), uniqid('entwechat', true));
-    }	
+    }
 
     /**
      * @param string $path
